service de estoque e funcionário

See merge request dthlogus/ohwaiterapi!2

+++ src/main/java/br/com/valhalla/ohwaiter/repository/EstoqueRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.valhalla.ohwaiter.model.Estoque;
 
 @Repository
-public interface EstoqueRepository extends JpaRepository<Estoque, Integer>{
+public interface EstoqueRepository extends JpaRepository<Estoque, Long>{
     
 }
